@@ -1,11 +1,11 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 29
     defaultConfig {
         applicationId "kr.co.aiotlab.www"
-        minSdkVersion 19
-        targetSdkVersion 28
+        minSdkVersion 21
+        targetSdkVersion 29
         versionCode 1
         versionName "1.0"
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
@@ -26,15 +26,15 @@ repositories {
 }
 
 dependencies {
-    implementation fileTree(include: ['*.jar'], dir: 'libs')
-    implementation 'com.android.support:appcompat-v7:28.0.0'
-    implementation 'com.android.support:mediarouter-v7:28.0.0'
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
 
-    implementation 'com.android.support:support-v4:28.0.0'
     // 카드뷰 사용을 위한 라이브러리
+    //noinspection GradleCompatible
     implementation 'com.android.support:cardview-v7:28.0.0'
     // 리사이클러뷰를 사용하기 위한 라이브러리
+    //noinspection GradleCompatible
     implementation 'com.android.support:recyclerview-v7:28.0.0'
+    //noinspection GradleCompatible
     implementation 'com.android.support:design:28.0.0'
 
     implementation 'com.android.support.constraint:constraint-layout:1.1.3'
@@ -56,5 +56,19 @@ dependencies {
     implementation('org.eclipse.paho:org.eclipse.paho.android.service:1.0.2') {
         exclude module: 'support-v4'
     }
+
+    //androidx
+    implementation 'androidx.recyclerview:recyclerview:1.1.0'
+    implementation 'androidx.constraintlayout:constraintlayout:2.0.0-beta4'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
+    implementation 'com.android.support.constraint:constraint-layout:2.0.0-beta4'
+    implementation "androidx.swiperefreshlayout:swiperefreshlayout:1.0.0"
+    implementation "com.google.android.material:material:1.0.0"
+
+    // date picker 라이브러리
+    implementation 'com.wdullaer:materialdatetimepicker:2.3.0'
+    // Glide 라이브러리
+    implementation 'com.github.bumptech.glide:glide:3.7.0'
+
 }
 apply plugin: 'com.google.gms.google-services'
