@@ -1,7 +1,7 @@
 package kr.co.aiotlab.www;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.util.Log;
 import android.widget.CalendarView;
 
