@@ -3,7 +3,7 @@ package kr.co.aiotlab.www.Main_UI;
 import android.content.Intent;
 import android.os.Bundle;
 import android.os.Handler;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 
 import kr.co.aiotlab.www.EnterIPandPort;
 import kr.co.aiotlab.www.R;
