@@ -1,23 +1,18 @@
 package kr.co.aiotlab.www.LineChart;
 
+import android.content.Context;
 import android.content.SharedPreferences;
 import android.graphics.Color;
 import android.os.Bundle;
 import android.os.Handler;
 import android.os.Looper;
-import android.preference.PreferenceManager;
-import android.support.annotation.NonNull;
-import android.support.v4.app.Fragment;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import com.github.mikephil.charting.charts.LineChart;
 import com.github.mikephil.charting.components.AxisBase;
-import com.github.mikephil.charting.components.LimitLine;
 import com.github.mikephil.charting.components.XAxis;
 import com.github.mikephil.charting.components.YAxis;
 import com.github.mikephil.charting.data.Entry;
@@ -37,214 +32,174 @@ import java.util.Collections;
 import java.util.Date;
 import java.util.Map;
 
+import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
 import kr.co.aiotlab.www.R;
 
-import static kr.co.aiotlab.www.Main_UI.BottomFourthFragment.spinner_month;
-import static kr.co.aiotlab.www.Main_UI.BottomFourthFragment.spinner_year;
+import static kr.co.aiotlab.www.Main_UI.BottomFourthFragment.txt_day;
 
-public class LinechartSensor1 extends Fragment implements View.OnClickListener {
-    public static final String TAG = "LINE1";
+public class Linechart_Power_Minute extends Fragment implements View.OnClickListener {
 
     private LineChart mChart;
-
     private View view;
     private FirebaseDatabase mDatabase;
     private Date date;
     private TextView txt_backChart, txt_nextChart;
-    private static TextView txt_chart_date;
     String time, year, month, day;
-    int set_temp;
     String yearData, monthData, dayData, hourData, minuteData;
-    DatabaseReference temp_minute_1Ref, ref;
+    DatabaseReference current1Ref;
     ChildEventListener childEventListener;
+    // 저장된 날짜 담아오기
+    private String yearString, monthString, dayString;
 
-    public static LinechartSensor1 newInstance() {
-        LinechartSensor1 f = new LinechartSensor1();
+    public static Linechart_Power_Minute newInstance() {
+        Linechart_Power_Minute f = new Linechart_Power_Minute();
         return f;
     }
 
     @Nullable
     @Override
     public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, Bundle savedInstanceState) {
-        view = inflater.inflate(R.layout.layout_sensor1, container, false);
-        mDatabase = FirebaseDatabase.getInstance();
-        mChart = view.findViewById(R.id.linechart_sensor1);
+        view = inflater.inflate(R.layout.layout_sensor7, container, false);
+
+        mChart = view.findViewById(R.id.linechart_sensor7);
         mChart.setNoDataText("데이터를 불러오는 중입니다.");
         mChart.setNoDataTextColor(Color.BLUE);
+        mDatabase = FirebaseDatabase.getInstance();
 
-        txt_chart_date = view.findViewById(R.id.txt_chart_date);
         txt_nextChart = view.findViewById(R.id.txt_nextChart);
-
         txt_backChart = view.findViewById(R.id.txt_backChart);
         txt_nextChart.setOnClickListener(this);
         txt_backChart.setOnClickListener(this);
 
-        SharedPreferences set_temp_pref = PreferenceManager.getDefaultSharedPreferences(getContext());
-        set_temp = Integer.parseInt(set_temp_pref.getString("SETTEMP", "0"));
-        LimitLine limit_high_risk = new LimitLine(set_temp, "설정 온도");
-
-        limit_high_risk.setLineColor(Color.RED);
-        limit_high_risk.setLineWidth(2f);
-        limit_high_risk.enableDashedLine(10f, 10f, 0f);
-        limit_high_risk.setLabelPosition(LimitLine.LimitLabelPosition.RIGHT_TOP);
-        limit_high_risk.setTextColor(Color.RED);
-
-        YAxis leftAxis = mChart.getAxisLeft();
-        leftAxis.removeAllLimitLines();
-        leftAxis.addLimitLine(limit_high_risk); //상한선 추가(왼쪽 축)
-        leftAxis.enableGridDashedLine(10f, 10f, 0f);
-        leftAxis.setDrawLimitLinesBehindData(true);
-
         getNowTime();
-        txt_chart_date.setText(spinner_year + "년 " + spinner_month + "월 " + day + "일");
+        SharedPreferences sharedPreferences = this.getActivity().getSharedPreferences("TimeSave", Context.MODE_PRIVATE);
+        yearString = sharedPreferences.getString("YEAR", year);
+        monthString = sharedPreferences.getString("MONTH", month);
+        dayString = sharedPreferences.getString("DAY", day);
+
         showLineChartData();
 
         return view;
     }
 
-
     @Override
     public void onClick(View v) {
-        temp_minute_1Ref.removeEventListener(childEventListener);
+        current1Ref.removeEventListener(childEventListener);
         switch (v.getId()) {
             case R.id.txt_backChart:
-                day = String.valueOf(Integer.parseInt(day) - 1);
-                if (Integer.parseInt(day) < 10) {
-                    day = "0" + day;
+                dayString = String.valueOf(Integer.parseInt(dayString) - 1);
+                if (Integer.parseInt(dayString) < 10) {
+                    dayString = "0" + dayString;
                 }
 
-                if (Integer.parseInt(day) <= 1) {
+                if (Integer.parseInt(dayString) <= 1) {
                     txt_backChart.setVisibility(View.INVISIBLE);
                 } else {
                     txt_nextChart.setVisibility(View.VISIBLE);
                 }
-
-                txt_chart_date.setText(spinner_year + "년 " + spinner_month + "월 " + day + "일");
-                try {
-                    showLineChartData();
-                } catch (IndexOutOfBoundsException e) {
-                    Toast.makeText(getContext(), "데이터가 없습니다.", Toast.LENGTH_SHORT).show();
-                }
-
-                Log.d(TAG, "onClick: day - 1 = " + day);
+// TextView 일자 변경
+                txt_day.setText(dayString);
+                // SharedPreference에 일자 저장
+                saveDateDay(dayString);
+                showLineChartData();
                 break;
             case R.id.txt_nextChart:
-                day = String.valueOf(Integer.parseInt(day) + 1);
-                if (Integer.parseInt(day) < 10) {
-                    day = "0" + day;
+                dayString = String.valueOf(Integer.parseInt(dayString) + 1);
+                if (Integer.parseInt(dayString) < 10) {
+                    dayString = "0" + dayString;
                 }
 
-                if (Integer.parseInt(day) > 30) {
+                if (Integer.parseInt(dayString) > 30) {
                     txt_nextChart.setVisibility(View.INVISIBLE);
                 } else {
 
                     txt_backChart.setVisibility(View.VISIBLE);
                 }
-
-                txt_chart_date.setText(spinner_year + "년 " + spinner_month + "월 " + day + "일");
-                try {
-                    showLineChartData();
-                } catch (NullPointerException e) {
-                    Toast.makeText(getContext(), "데이터가 없습니다.", Toast.LENGTH_SHORT).show();
-                }
-                Log.d(TAG, "onClick: day + 1 = " + day);
+// TextView 일자 변경
+                txt_day.setText(dayString);
+                // SharedPreference에 일자 저장
+                saveDateDay(dayString);
+                showLineChartData();
                 break;
         }
     }
-
-
+    private void saveDateDay(String dayString) {
+        SharedPreferences sharedPreferences = this.getActivity().getSharedPreferences("TimeSave", Context.MODE_PRIVATE);
+        SharedPreferences.Editor editor = sharedPreferences.edit();
+        if (Integer.parseInt(yearString) < 10) {
+            editor.putString("DAY", "0" + dayString);
+        } else {
+            editor.putString("DAY", dayString);
+        }
+        editor.apply();
+    }
     private void getNowTime() {
         //현재 시간
         long now = System.currentTimeMillis();
         date = new Date(now);
-
         SimpleDateFormat sdfYear = new SimpleDateFormat("yyyy");
         SimpleDateFormat sdfMonth = new SimpleDateFormat("MM");
         SimpleDateFormat sdfDay = new SimpleDateFormat("dd");
         SimpleDateFormat sdfHour = new SimpleDateFormat("HH");
         SimpleDateFormat sdfMinute = new SimpleDateFormat("mm");
-
-
         yearData = sdfYear.format(date);
         monthData = sdfMonth.format(date);
         dayData = sdfDay.format(date);
         hourData = sdfHour.format(date);
         minuteData = sdfMinute.format(date);
-
         year = yearData;
         month = monthData;
         day = dayData;
     }
 
     private void showLineChartData() {
-
         new Thread(new Runnable() {
-            final ArrayList<Float> temp_data_minute = new ArrayList<>();
-            final ArrayList<String> xEntrys = new ArrayList<>();
-
             @Override
             public void run() {
-                // 데이터 받아오기, 갱신
-                // Log.d(TAG, "MQTT_Temperature_minute_" + year + month + day);
-                temp_minute_1Ref = mDatabase.getReference("MQTT_Temperature_minute_" + spinner_year + spinner_month + day);
+                current1Ref = mDatabase.getReference("Current_minute_" + yearString + monthString + day);
+                final ArrayList<Float> current1_data = new ArrayList<>();
+                final ArrayList<String> xEntry = new ArrayList<>();
+
                 childEventListener = new ChildEventListener() {
                     @Override
                     public void onChildAdded(@NonNull DataSnapshot dataSnapshot, String s) {
                         try {
+
                             Map<String, Object> data = (Map<String, Object>) dataSnapshot.getValue();
 
-                            //Log.d(TAG, "onChildAdded:" + data.get("temperature").toString());
-                            //Log.d(TAG, "onChildAdded:" + data.get("timestamp").toString());
+                            float current1 = Float.parseFloat(data.get("Current1").toString());
+                            String timestamp = data.get("timestamp").toString();
+                            String timestamp_time = timestamp.substring(8, 10) + ":" + timestamp.substring(10, 12);
+                            xEntry.add(timestamp_time);
 
+                            current1_data.add(current1);
 
-                            float temp_minute = Float.parseFloat(data.get("temperature").toString());
-                            String timestamp = (data.get("timestamp").toString());
-
-                            String timestamp_hour = timestamp.substring(8, 10);
-                            String timestamp_minute = timestamp.substring(10, 12);
-                            String xAxis_time = timestamp_hour + ":" + timestamp_minute;
-
-                            temp_data_minute.add(temp_minute);
-                            xEntrys.add(xAxis_time);
-
-                            //Log.d(TAG, "xaxisSize: " + xEntrys.size());
-                            //Log.d(TAG, "yaxisSize: " + watt_data.size());
-
-                            final String[] xaxes = new String[xEntrys.size()];
-
-                            final ArrayList<Entry> yData = new ArrayList<>();
-
+                            ArrayList<Entry> yData = new ArrayList<>();
                             ArrayList<ILineDataSet> dataSets = new ArrayList<>();
 
                             float total = 0;
                             //a의 data로 a는 value값을 int형으로 변환한 것. value는 데이터가 바뀔 때의 값으로 업데이트 될 때마다 값이 나온다.
-                            for (int j = 0; j < temp_data_minute.size(); j++) {
-                                yData.add(new Entry(j, temp_data_minute.get(j)));
-                                xaxes[j] = xEntrys.get(j);
-                                // 평균 사용량
-                                total = total + temp_data_minute.get(j);
+                            for (int j = 0; j < current1_data.size(); j++) {
+                                yData.add(new Entry(j, current1_data.get(j) * 220));
+
+                                total = total + current1_data.get(j);
                             }
 
-                            float average = total / temp_data_minute.size();
+                            float average = total / current1_data.size() * 220;
 
-                            //Log.d(TAG, "Ysize:" + yData.size());
-                            //Log.d(TAG, "Xsize:" + xaxes.length);
-
-                            //Line data setting
-                            LineDataSet set1 = new LineDataSet(yData, "온도 (℃)," + "  평균 : " + average + "℃");
+                            LineDataSet set1 = new LineDataSet(yData, "평균 전력 = " + average + "W");
                             set1.setFillAlpha(110);
                             set1.setColor(Color.BLACK);
                             set1.setLineWidth(3f);
+
 
                             dataSets.add(set1);
 
                             LineData data2 = new LineData(dataSets);
 
-                            //mChart.setVisibleXRange(10, 20);
-                            //mChart.moveViewToX(watt_data.size());
                             mChart.animateX(1000);
-
                             //      mChart.setOnChartGestureListener(SecondFragmentThermo.this);
                             //      mChart.setOnChartValueSelectedListener(SecondFragmentThermo.this);
 
@@ -257,40 +212,31 @@ public class LinechartSensor1 extends Fragment implements View.OnClickListener {
                             mChart.getAxisRight().setEnabled(false);
 
                             XAxis xAxis = mChart.getXAxis();
-                            xAxis.setGranularity(1f);
                             xAxis.setValueFormatter(new IAxisValueFormatter() {
                                 @Override
                                 public String getFormattedValue(float value, AxisBase axis) {
                                     String val = null;
                                     try {
-                                        val = xEntrys.get((int) value);
+                                        val = xEntry.get((int) value);
                                     } catch (IndexOutOfBoundsException e) {
                                         axis.setGranularityEnabled(false);
                                     }
                                     return val;
                                 }
                             });
-
+                            xAxis.setGranularity(1);
                             xAxis.setPosition(XAxis.XAxisPosition.BOTTOM);
-                /*xAxis.setValueFormatter(new IAxisValueFormatter() {
-                   @Override
-                   public String getFormattedValue(float value, AxisBase axis) {
-                       return xaxes[(int) value];
-                   }
-               }); */
 
                             YAxis leftAxis = mChart.getAxisLeft();
 
-                            float y_max = Collections.max(temp_data_minute);
-                            float y_min = Collections.min(temp_data_minute);
+                            //측정 값중 최댓값 구하기
+                            float y_max = Collections.max(current1_data);
 
-                            leftAxis.setAxisMaximum(y_max + 3);               //보여지는 최대
-                            leftAxis.setAxisMinimum(y_min - 3);
-
+                            leftAxis.setAxisMaximum(y_max + 50);              //보여지는 최대
+                            leftAxis.setAxisMinimum(0f);
                         } catch (Exception e) {
-                            e.printStackTrace();
-                        }
 
+                        }
                     }
 
                     @Override
@@ -304,7 +250,7 @@ public class LinechartSensor1 extends Fragment implements View.OnClickListener {
                     }
 
                     @Override
-                    public void onChildMoved(@NonNull DataSnapshot dataSnapshot, @android.support.annotation.Nullable String s) {
+                    public void onChildMoved(@NonNull DataSnapshot dataSnapshot, String s) {
 
                     }
 
@@ -313,12 +259,11 @@ public class LinechartSensor1 extends Fragment implements View.OnClickListener {
 
                     }
                 };
-
                 Handler handler = new Handler(Looper.getMainLooper());
                 handler.post(new Runnable() {
                     @Override
                     public void run() {
-                        temp_minute_1Ref.addChildEventListener(childEventListener);
+                        current1Ref.addChildEventListener(childEventListener);
 
                     }
                 });
@@ -327,5 +272,4 @@ public class LinechartSensor1 extends Fragment implements View.OnClickListener {
 
 
     }
-
 }
